feat: Add plain version algorithm
The plain version computes the SSE element by element in code then
relay on compiler to utilize AVX/SIMD to vectorize the computation.
The result is checked compare to the Matlab code.
The computation speed seems to be 300 times slower than sliding
window algorithm for single kernel.
The debug tool addreddsanitizer is enabled, has to be removed for
final speed test.

Now the plain version and the sliding window version share the
same class name. And in main.cpp file, depending on the marco
flag given by the compiler, main.cpp would include different
hpp file.
Then the makefile defines two compile task with different flag.
Therefore, we can have two binary files with different algorithm.
This may not be the best approach (what if we have .cpp file for
the .hpp to link?). But we still avoid repeating some of the code
and overhead of vitural function.
`make all` then `./runtest.sh` to compile and run both algorithms.