feat: Multi kernel classification
Now the program would output the ID of the best match motion.
The result is checked compare to Matlab code.
Also modified the CSV file header to contain ID and name so that
the classification is consistent with Matlab result.
Now the new format data and code is not compatible with last commit.

Note: <known bug> I am breaking the rule of 3/5/0: MotionMatcher.hpp
only has a custom destructor but no copy-constructor nor assignment
operator. Therefore, the code is under the risk of undefined behavior
if we copied it, which happens in the vector container. I avoided copy
by reserve the memory space of vector before push_back/emplace_back
object into vector. Ideally, I would use unique_ptr to address this
properly, but for now, it should not effect the result and speed.